Rename admin flag page to flags
